Allow bootloader to be updated from interface
Add support for updating the bootloader from the interface.  The first
operation performed when updating the bootloader is to erase the first
sector of the bootloader and write the interface's vector table to it.
That way if power is lost during the update the device will still be in
a usable state, and another bootloader update can be attempted.

As the last step in updating the bootloader, the CRC of the bootloader
is checked.  If the CRC is correct, the first sector of the bootloader
is erased and updated with the bootloaders vector table.  Otherwise
the update is aborted and marked as a failure.  This is to prevent a
corrupt bootloader image from leaving devices in an unusable state.