Fix USB hang on lpc11u35 some MSD transfers
When a MSD IN and OUT endpoint data interrupt are pending at the
same time the OUT endpoint gets processed first, even if data arrived
on the IN endpoint first.  If mass storage is in the MSC_BS_CSW state
then the OUT packet will get ignored.  When the IN packet gets
processed the MSD state will change to MSC_BS_CBW.  Mass storage
then stops working because the device is waiting for an OUT packet
for the next transfer while the PC is waiting for an IN response.
On 20-60s after being in this state Windows will perform a USB reset
on this device, allowing MSD to continue.

This patch fixes the problem by processing endpoint interrupts from
highest index to lowest index.  This causes the IN endpoints to get
processed before the OUT endpoints.  This patch also adds an assert
which is triggered if this happens.

This problem does not effect the K20DX or KL26z since the endpoint
interrupt register is actually a FIFO where only 1 bit is set at a time.
The atsam3u is not effected since all the IN endpoints are processed
before the OUT endpoints.  If the order of endpoints is changed in
usb_config.c on the atsam3u then this same problem can be seen.

The ideal solution for this problem would be to have the MSD driver
only enable IN or OUT at any given time and not both.  This is not
possible with the current USB driver because reading data from an
OUT endpoint (USBD_ReadEP) prepares that same OUT endpoint for more
data.  This problem does not apply to an IN endpoint which is
inactive by default, and is only activated by writing data to it.
To overcome this always process the IN endpoint first, since there
was an action which intentionally enabled it.