Fix serial test on OSX by keeping port open
When a serial port OSX is closed a beak command is sent.  This causes
the serial test to fail.  This patch updates serial_test.py to change
the baudrate by using the 'baudrate' property rather than by closing
the serial port and opening a new instance at the proper speed.  This
fixes the serial port test.