Fix assert programming certain nrf51 hex files
Some hex files built for the nrf51 program both the UICR anf FICR in
addition to regular flash.  This triggers an assert because  this is
not in program flash, which target_flash_erase_sector_size checks for.
This patch fixes that assert by always returning a sector size, even
if it is outside program flash.