Allow interface to dump target information
Add the ability to enter dump mode by copying DUMP.ACT on to the
filesystem.  When in dump mode the target is halted and the extra
files ROM.BIN, RAM.BIN, REGS.TXT and RESUME.TXT are present.
To exit dump mode and resume target execution delete the file
RESUME.TXT.