Fix reported FS corruption after file creation
When a file is created the root directory is updated but the FAT
table and actual file contents are not.  If chkdsk is run in this
state it will report that the filesystem is corrupted.  This patch
fixes that problem by making filesystem reads always return the exact
same filesystem image.